Added full MC to allow modification of cloud-init
In a future release of MC it might be possible to add packages to be installed via puppet hieradata. This would allow csh to be installed directly on the nodes rather than using the csh version installed in cvmfs. The main reason this is helpful is that /bin/csh will exist, as it doesn't right now. A particular bit of software that is used in the workshop looks for /bin/csh specifically. Right now I am going ot just create a symbolic link with cloud-init to the cvmfs version.